pyproject.toml
[project]
name = "business-assistant-calendar-plugin"
version = "0.1.0"
description = "Google Calendar plugin for Business Assistant v2"
requires-python = ">=3.11,<3.13"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.2.0",
    "python-dateutil>=2.8.0",
    "icalendar>=5.0.0",
    "business-assistant-v2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/business_assistant_calendar"]

[tool.uv.sources]
business-assistant-v2 = { path = "D:/GIT/BenjaminKobjolke/business-assistant-v2", editable = true }

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
